+++ java/src/main/java/com/techelevator/dao/CommentsDao.java
@@ -5,7 +5,7 @@ import com.techelevator.model.Comments;
 import java.util.List;
 
 public interface CommentsDao {
-    List<Comments> getCommentsByPhotoId(Long id);
+    List<Comments> getCommentsByPhotoId(int id);
 
     int createComment(Comments comments, String username);
